+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 0.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Allow to specify height and width for barcode generation
+  [laulaz]
 
 
 0.1 (2015-11-12)